Extract addReducerEnhancer into a util
