+++ client/state/utils/index.ts
@@ -8,3 +8,4 @@ export { withEnhancers } from './with-enhancers';
 export { withoutPersistence } from './without-persistence';
 export { withStorageKey } from './with-storage-key';
 export { addReducer, combineReducers } from './reducer-utils';
+export { addReducerEnhancer } from './add-reducer-enhancer';
